[nrfconnect] Expose FactoryDataBase outside nrf implementation
- Moved FactoryDataProviderBase outside the FactoryDataProvider.h
to not build the FactoryDataProvider.h while a custom
implementation is used.
- Added Kconfig choice to select the factory data backend to
easily switch between NRFCONNECT and CUSTOM ones.
The new approach is backward compatible, and by default it is set
to CHIP_FACTORY_DATA_NRFCONNECT_BACKEND.

To use custom backend:

Set CONFIG_CHIP_FACTORY_DATA_CUSTOM_BACKEND=y and
SB_CONFIG_MATTER_FACTORY_DATA_GENERATE=n
