+++ src/platform/nrfconnect/BUILD.gn
@@ -47,6 +47,7 @@ static_library("nrfconnect") {
     "DiagnosticDataProviderImplNrf.cpp",
     "DiagnosticDataProviderImplNrf.h",
     "ExternalFlashManager.h",
+    "FactoryDataProviderBase.h",
     "FactoryResetTestEventTriggerHandler.cpp",
     "FactoryResetTestEventTriggerHandler.h",
     "InetPlatformConfig.h",
