+++ src/test_driver/nrfconnect/prj.conf
@@ -90,6 +90,7 @@ CONFIG_CHIP_PROJECT_CONFIG="main/include/CHIPProjectConfig.h"
 # Don't erase all settings as it deinitializes NVS/ZMS and causes issues with
 # testing Server::ScheduleFactoryReset
 CONFIG_CHIP_FACTORY_RESET_ERASE_SETTINGS=n
+CONFIG_CHIP_FACTORY_DATA_NONE=y
 
 # Use default ICD poll intervals to keep compatibility with tests
 CONFIG_CHIP_ICD_SLOW_POLL_INTERVAL=5000
